+++ ospray/render/Renderer.cpp
@@ -20,6 +20,7 @@ static FactoryMap<Renderer> g_renderersMap;
 Renderer::Renderer()
 {
   managedObjectType = OSP_RENDERER;
+  pixelFilter = nullptr;
 }
 
 std::string Renderer::toString() const
@@ -53,6 +54,8 @@ void Renderer::commit()
 
   materialData = getParamDataT<Material *>("material");
 
+  setupPixelFilter();
+
   if (materialData)
     ispcMaterialPtrs = createArrayOfIE(*materialData);
   else
@@ -67,7 +70,8 @@ void Renderer::commit()
         backplate ? backplate->getIE() : nullptr,
         ispcMaterialPtrs.size(),
         ispcMaterialPtrs.data(),
-        maxDepthTexture ? maxDepthTexture->getIE() : nullptr);
+        maxDepthTexture ? maxDepthTexture->getIE() : nullptr,
+        pixelFilter ? pixelFilter->getIE() : nullptr);
   }
 }
 
@@ -140,6 +144,40 @@ OSPPickResult Renderer::pick(
   }
 
   return res;
+}
+
+void Renderer::setupPixelFilter()
+{
+  OSPPixelFilterTypes pixelFilterType = (OSPPixelFilterTypes)getParam<int>("pixelFilter", OSPPixelFilterTypes::OSP_PIXELFILTER_GAUSS);
+  pixelFilter = nullptr;
+  switch(pixelFilterType){
+    case OSPPixelFilterTypes::OSP_PIXELFILTER_BOX:
+    {
+        pixelFilter = ospcommon::make_unique<ospray::BoxPixelFilter>();
+        break;
+    }
+    case OSPPixelFilterTypes::OSP_PIXELFILTER_BLACKMAN_HARRIS:
+    {
+        pixelFilter = ospcommon::make_unique<ospray::BlackmanHarrisLUTPixelFilter>();
+        break;
+    }
+    case OSPPixelFilterTypes::OSP_PIXELFILTER_MITCHELL:
+    {
+        pixelFilter = ospcommon::make_unique<ospray::MitchellNetravaliLUTPixelFilter>();
+        break;
+    }
+    case OSPPixelFilterTypes::OSP_PIXELFILTER_POINT:
+    {
+        pixelFilter = ospcommon::make_unique<ospray::PointPixelFilter>();
+        break;
+    }
+    case OSPPixelFilterTypes::OSP_PIXELFILTER_GAUSS:
+    default:
+    {
+        pixelFilter = ospcommon::make_unique<ospray::GaussianLUTPixelFilter>();
+        break;
+    }
+  }
 }
 
 OSPTYPEFOR_DEFINITION(Renderer *);